pyproject.toml
[project]
name = "exif-finder"
version = "0.0.2"
description = "Command-line tool to find JPEG images by EXIF date"
authors = [{name = "Example", email = "example@example.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0"
]

[project.scripts]
exif-finder = "exif_finder.__main__:main"
